dbt_serverless/requirements.txt
fastapi>=0.88.0,<0.89.0
pydantic>=1.9.2,<2.0.0
uvicorn>=0.18.3,<0.19.0
pyyaml>=6.0,<6.1
dbt-bigquery>=1.3.0,<1.4.0
google-cloud-storage>=2.7.0,<2.8.0
